ChatGPT/Voice Assistant/requirements.txt
pydub
SpeechRecognition
openai==0.27.4
openai-whisper
numpy
gTTS
click
pyaudio
python-dotenv